Remove wait from escape key in tmux
Tmux was inserting a 500ms wait before sending the keycode for escape.
This was noticeably a problem in neovim and so I have reduced it down to
10ms.

So this issues isn't present in Vim because of the way neovim handles
escape. Neovim handles alt codes, which the escape character is one of.
It is these alt codes that tmux is hanging onto.