Replace denite with fzf
Fzf has a much cleaner interface and most of the configuration is
already done for me. I also like that the completion is available on the
command line, having a consistent interface between both it and vim.
Additionally I can find more useful documentation on using fzf than
deoplete, so it seems better supported.