Modify git aliases to allow for golang
I have been using `go`, `gs` and a few other aliases, however these are
colliding with other programs like golang's `go` binary and
`ghostscript`. To alleviate some of this and have a more cohesive
interface for git I am introducing a `g` alias for git and then using
the git aliases for the rest of the commands.