Function to check function/directory/file exists
Rather than writing the check for existence for each of the files I am
loading into the zsh configuration, I have written a series of functions
which perform the check, and make the required modification. I have
included functions for the modification of the PATH variable, to source
a file, and to alias a function.