Attempt to set TERM appropriately
When tmux > 2 is installed, it also installs a tmux-256color terminal
which is understood. However versions of tmux prior to this don't
have this feature which can cause issues in the translation of commands
to actions when ssh'd into a machine while in a tmux session. This
attempts to fix the issue by conditionally setting the TERM variable
when the issue arises.