Cleanup fish configuration
This moves the initialisation of the path to autoload in the conf.d
folder as a result of using the pyenv plugin for loading pyenv into the
shell. This change significantly reduces the startup time of the shell,
getting it down to 0.06s (about 5x faster than zsh).

Additionally this removes a lot of the functions/imports I don't really
use in this updated workflow.