Configure ale as linting engine
I have previously been using neomake for my linting, however I have been
unhappy with it of late and the lack of support for textlint was a bit
of an issue. I have heard very good things about ale and with the
support for the plugins I wanted I decided to try ale out and see what I
think.