Modify PATH over fish_user_paths
Updating the fish_user_paths variable was causing some unusual problems
when dealing with pyenv shims. Often the order of PATH would be messed
up since pyenv edited the PATH directly, and the fish_user_paths
prepends the path resulting in the pyenv shims being halfway down the
list of paths to search.

This made it even more difficult to work out which version of things I
was trying to use.