Include vim-vinegar plugin once again
This plugin has a lot of useful features for dealing with the directory
structure within vim. There are still some things that nerdtree does
really well with listing out the directory structure, however with
vinegar being lightweight and quick to interact with there is a
significant benefit to using it.