ENH [vim] Create goyo configuration
Goyo is a plugin that removes many of the distractions from vim,
especially those that make writing code or editing simpler. The idea is
that I will be able to just write more effectively, not having to worry
about what words are put on the page, instead focus on just getting them
there.