[tmux] Configure tm utility properly
This fixes the issue o not breaking out of the select properly, using
`break 2`, i.e. passing the argument 2 to break to break out of multiple
levels. Additionally the `;;` are not part of the break statement,
instead they signify the end of the case options, so I have semantically
signified that.