[vim] Add vimpyter plugin for editing jupyter files
The default vimpyter plugin uses notedown for the conversion which I
have found to not be to my liking. A nicer tool for the conversion of
jupyter files is jupytext. So the vimpyter plugin is my own spin, which
uses juptyext for the conversion of jupyter files.

It is likely that I will create my own vim-jupytext plugin which handles
the conversion in a manner that is more cohesive with the operation of
jupytext.