[vim] Optimise pandoc configuration
This makes a few changes to the pandoc configuration, improving the
interface. Most noticeable is the syntax highlighting of embedded python
code blocks. Currently this is only python since that is what I write,
although that is reasonably likely to expand.

Additionally I have disabled the formatting module since I have now
decided to use semantic newlines. The pandoc formatting was making it
difficult to perform the formatting when adding a newline at the start
of a word.