[vim] Cycle through hunks in all files
This remaps the ]c and [c key bindings to move through the changes in
all the files of a repository, not just the current file. The idea is
that this is a better approach to running the `git add -p` command since
I don't have to leave vim and have more context for the changes.