[vim] Support devicons in fzf preview
This adds icons to each file of the preview window in fzf. While this is
a really cool idea, it requires a NerdFont or a NerdFont as fallback.
Since alacrity doesn't support fallbacks properly, and the NerdFont
Iosevka doesn't work for me, particularly with italics, I will save this
for later.