[vim] Revert to using C-hjkl instead of A-hjkl
The alt mappings were too similar to the i3 bindings and had the
additional complication of being in different positions on different
keyboards. Additionally there is the problem of having to move too far
to get to the bindings. The main issue I had previously was the
remapping of the C-h backspace, however that isn't as much as a
problem as I am making it out to be.