nvim: Set undofile to true
When the undofile is turned on, vim saves a persistent history of
changes to a file when writing to a buffer and restores that state when
reading the buffer. This allows for the undo history of a file to
transcend a single editing session.

This behaviour is what I normally intend when editing a file,
particularly when quickly opening and closing a neovim session.