nvim: Replace vimwiki with corpus
Vimwiki is a complex plugin that contains a whole lot of features. This
large range of features includes a collection of tools that I am
fighting against, including the many modifications to shortcuts and
redefinition of key commands. The key issue I have is vimwiki
overwriting the filetype to vimwiki, rather than leveraging the markdown
filetype with all the plugins and tools I am using in it.

The plugin, Corpus, is a more lightweight plugin that contains the key
components of vimwiki that I want to use. Primarily the management of
wikis. The really cool feature of Corpus is the search functionality,
making finding information far simpler. There is some functionality that
I am not overly sure about is the diary, however this is definitely
something simple I can add if necessary.