nvim: Switch to dirvish over vinegar
The dirvish plugin is simpler and has fewer features, which is what I am
looking for here. It is also significantly faster and maps better to the
vim interface. This still retains what I like about vinegar, being the
easy access to files, without some of the weirdness of netrw.