nvim: Include installation for vim-plug
This includes the steps for installing vim-plug automatically should it
not already be installed. This makes the setup on new machines
significantly easier.

I don't particularly like this solution, however this functionality is
required now, rather than waiting for the ideal solution. I do want to
explore the possibility of using alternative tools for managing plugins.