vim: Update python configuration including comments
This updates the ftplugin configuration for python files. The main
change is the removal of set spell. This is because with the update to
treesitter parsing, there is no longer the definition of spell regions
within python, making the spell check no longer useful.

Additionally this provides some additional comments on the configuration
options that remain included in the ftplugin file.