nvim: Convert from vim-plug to packer.nvim
Using packer.nvim, I am able to move more of my configuration to Lua, as
packer does all the configuration in lua. This is an excellent
advantage in the movement of my configuration to lua.

An additional advantage of using packer is that it provides a method of
specifying dependent packages. The dependency between plugins is
something that is becoming more and more common within the neovim
ecosystem, particularly with the move to lua for plugins. There are some
fundamental plugins that handle a lot of the hard work making plugins
easier to write.

The initial installation of packer requires a little work to get setup,
and some more in depth work to keep updated. However the benefits of
this plugin manager outweigh the additional complexities.